Janitor & Summaries
The goal of this PR is to provide a new summaries table that holds the number
of registrants, waitlisted people, etc. A cronjob-style CloudWatch event triggers
a Janitor Lambda to keep this table up to date.

Once this table is populated, we will be able to build on top of this work to add
automatic waitlist processing.